requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
google-auth==2.29.0
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
repoze.lru==0.7
requests==2.31.0
Routes==2.5.1
rsa==4.9
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
